+++ src/core/section.h
@@ -104,7 +104,8 @@ namespace PixelMaestro {
 					 * Divide the x interval by the Maestro's refresh rate, then divide the Section's x-axis size by the result.
 					 * This gives you the number of pixels to move over per refresh.
 					 */
-					float x = dimensions->x / (float)(interval_x / (float)refresh_interval);
+					float x = 0;
+					if (interval_x != 0.0f) x = dimensions->x / (float)(interval_x / (float)refresh_interval);
 					// If x is less than 1 pixel, calculate the amount of time until the Section scrolls by 1 pixel.
 					if (x > 0 && x < 1) {
 						uint16_t interval = (1 / x) * refresh_interval;
@@ -121,7 +122,8 @@ namespace PixelMaestro {
 						step_x = x;
 					}
 
-					float y = dimensions->y / (float)(interval_y / (float)refresh_interval);
+					float y = 0;
+					if (interval_y != 0.0f) dimensions->y / (float)(interval_y / (float)refresh_interval);
 					if (y > 0 && y < 1) {
 						uint16_t interval = (1 / y) * refresh_interval;
 						if (timer_y) {